+++ src/neat/Pool.java
@@ -22,7 +22,7 @@ public class Pool implements Serializable {
     public int currentSpecies;
     public int currentGenome;
     public int currentFrame;
-    public int maxFitness = Integer.MIN_VALUE;
+    public long maxFitness = Integer.MIN_VALUE;
     public static final double Population = 300;
     public static final double DeltaDisjoint = 2.0;
     public static final double DeltaWeights = 0.4;
@@ -68,7 +68,7 @@ public class Pool implements Serializable {
 
     public Genome getbest() {
         Genome temp = this.Species.get(0).Genomes.get(0);
-        int maxfit = Integer.MIN_VALUE;// temp.fitness;
+        long maxfit = Integer.MIN_VALUE;// temp.fitness;
         for (Species s : this.Species) {
             for (Genome g : s.Genomes) {
                 if (g.getFitness() != 0) {
@@ -86,9 +86,9 @@ public class Pool implements Serializable {
         return temp;
     }
 
-    public int getTopfitness() {
+    public long getTopfitness() {
         // Genome temp = this.Species.get(0).Genomes.get(0);//this.Species.get(0).Genomes.get(0);
-        int maxfit = Integer.MIN_VALUE;// temp.fitness;
+        long maxfit = Integer.MIN_VALUE;// temp.fitness;
         for (Species s : this.Species) {
             for (Genome g : s.Genomes) {
                 if (g.getFitness() != 0) {
@@ -432,7 +432,7 @@ public class Pool implements Serializable {
             out.writeInt(this.Outputs);
             out.writeInt(this.generation);
             out.writeInt(this.Innovation);
-            out.writeInt(this.maxFitness);
+            out.writeLong(this.maxFitness);
             out.writeInt(this.Species.size());
             for (Species sp : this.Species) {
                 out.writeInt(sp.Genomes.size());
@@ -488,7 +488,7 @@ public class Pool implements Serializable {
             p.Outputs = out;
             p.generation = in.readInt();
             p.Innovation = in.readInt();
-            p.maxFitness = in.readInt();
+            p.maxFitness = in.readLong();
             p.currentGenome = 1;
             p.currentSpecies = 1;
             int Speciessize = in.readInt();
